[AIRLFOW-YYY] Move flower command to separate file
